Workspace restructure: libperl-proto0, libperl-macros, plan docs
Begin Step 0 of the libperl-macrogen-driven rebuild plan
(see new `docs/plan/README.md`).

Move the existing prototype API into a new workspace member
`libperl-proto0`, create an empty `libperl-macros` proc-macro crate,
and slim the top-level `libperl-rs` to a near-empty re-export of
`libperl-sys`.

Workspace now:

    libperl-rs/                 top-level (new API, almost empty stub)
    ├── Cargo.toml              workspace = [libperl-sys, libperl-config,
    │                                        libperl-proto0, libperl-macros]
    ├── src/lib.rs              `pub use libperl_sys;` only
    ├── docs/plan/              design plan and proc-macro feasibility note
    ├── libperl-config/         unchanged
    ├── libperl-sys/            unchanged
    ├── libperl-macros/         NEW: proc-macro = true,
    │                                build.rs sets `cfg(perl_useithreads)`,
    │                                src/lib.rs has a stub `#[thx]` attribute
    └── libperl-proto0/         NEW: frozen historical snapshot
        ├── Cargo.toml          publish = false, libperl-sys path-dep
        ├── build.rs            useithreads + perlapi_ver* cfg
        ├── src/{lib.rs, perl.rs}  moved from old top-level src/
        └── examples/           moved 13 examples + eg/ helpers
                                (libperl_rs::* → libperl_proto0::*)

After this commit:
- `cargo build --all` is green
- `cargo build --all --examples` is green
- `cargo run --example 000_perl_parse` (under proto0) runs
- `cargo run --example 100_scan_ops -- -le 'print "ok"'` walks the
  op tree to completion (thanks to the previous PL_op_name fix)

The `#[thx]` proc-macro is intentionally a no-op stub for now; the
real transformation is deferred to Step 1/3 of the plan.
