+++ libperl-macros/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "libperl-macros"
+version = "0.3.1"
+authors = ["Kobayasi, Hiroaki <buribullet@gmail.com>"]
+edition = "2024"
+description = "Procedural macros for libperl-rs (#[thx], #[xs_sub], xs_boot!)"
+
+repository = "https://github.com/hkoba/libperl-rs.git"
+license = "GPL-1.0-or-later OR Artistic-1.0-Perl"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = { version = "2", features = ["full"] }
+quote = "1"
+proc-macro2 = "1"
+
+[build-dependencies]
+libperl-config = { path = "../libperl-config", version = "^0.3.1" }
