+++ libperl-proto0/examples/000_perl_parse.rs
@@ -1,4 +1,4 @@
-use libperl_rs::Perl;
+use libperl_proto0::Perl;
 #[allow(unused)]
 use libperl_sys;
