+++ libperl-proto0/examples/001_perl_parse_args.rs
@@ -1,4 +1,4 @@
-use libperl_rs::Perl;
+use libperl_proto0::Perl;
 #[allow(unused)]
 use libperl_sys;
