+++ libperl-proto0/examples/103_scan_op_tree.rs
@@ -4,7 +4,7 @@
 use std::env;
 use std::ffi::CStr;
 
-use libperl_rs::Perl;
+use libperl_proto0::Perl;
 use libperl_sys::{OPf_KIDS, PL_op_name, op, unop};
 
 mod eg;
