+++ libperl-proto0/examples/109_scan_subs_intro1.rs
@@ -2,7 +2,7 @@
 use std::env;
 
 #[cfg(perlapi_ver26)]
-use libperl_rs::*;
+use libperl_proto0::*;
 
 #[cfg(perlapi_ver26)]
 use libperl_sys::opcode;
